Add OSLog support without poisoning namespace
This is an attempt to add support for OSLog without increasing the minimum platform requirements.

The use of a separate target means that the extra types required should not interfere with consumers.

The main question is whether `#if canImport(os)` will resolve to `false` on linux. If not then it may fail to compile and a platform check could be required.